+++ code/qcommon/cm_test.c
@@ -315,8 +315,8 @@ PVS
 */
 
 byte	*CM_ClusterPVS (qint cluster) {
-	if (cluster < 0 || cluster >= cm.numClusters || !cm.vised ) {
-		return cm.visibility;
+	if (cluster < 0 || cluster >= cm.numClusters || !cm.visibility ) {
+		return cm.novis;
 	}
 
 	return cm.visibility + cluster * cm.clusterBytes;