+++ code/qcommon/common.c
@@ -1210,7 +1210,7 @@ typedef struct zonedebug_s {
 
 typedef struct memblock_s {
 	struct memblock_s	*next, *prev;
-	qint			size;	// including the header and possibly tiny fragments
+	size_t			size;	// including the header and possibly tiny fragments
 	memtag_t	tag;	// a tag of 0 is a free block
 	qint			id;		// should be ZONEID
 #if defined(ZONE_DEBUG)
@@ -1224,8 +1224,8 @@ typedef struct freeblock_s {
 } freeblock_t;
 
 typedef struct memzone_s {
-	qint		size;			// total bytes malloced, including header
-	qint		used;			// total bytes used
+	size_t		size;		// total bytes malloced, including header
+	size_t		used;		// total bytes used
 	memblock_t	blocklist;	// start / end cap for linked list
 #if defined(USE_MULTI_SEGMENT)
 	memblock_t	dummy0;		// just to allocate some space before freelist
@@ -1305,7 +1305,7 @@ static void InsertFree( memzone_t *zone, memblock_t *block )
 
 #if defined(ZONE_DEBUG)
 	if ( block->size < sizeof( *fb ) + sizeof( *block ) ) {
-		Com_Error( ERR_FATAL, "InsertFree: bad block size: %i\n", block->size );
+		Com_Error( ERR_FATAL, "InsertFree: bad block size: %"PRIz"u\n", block->size );
 	}
 #endif
 
@@ -1327,11 +1327,11 @@ Separator is needed to avoid additional runtime checks in Z_Free()
 to prevent merging it with previous free block
 ================
 */
-static freeblock_t *NewBlock( memzone_t *zone, qint size )
+static freeblock_t *NewBlock( memzone_t *zone, size_t size )
 {
 	memblock_t *prev, *next;
 	memblock_t *block, *sep;
-	qint alloc_size;
+	size_t alloc_size;
 
 	// zone->prev is pointing on last block in the list
 	prev = zone->blocklist.prev;
@@ -1343,7 +1343,7 @@ static freeblock_t *NewBlock( memzone_t *zone, qint size )
 
 	sep = (memblock_t *) calloc( alloc_size, 1 );
 	if ( sep == NULL ) {
-		Com_Error( ERR_FATAL, "Z_Malloc: failed on allocation of %i bytes from the %s zone",
+		Com_Error( ERR_FATAL, "Z_Malloc: failed on allocation of %"PRIz"u bytes from the %s zone",
 			size, zone == smallzone ? "small" : "main" );
 		return NULL;
 	}
@@ -1379,7 +1379,7 @@ static freeblock_t *NewBlock( memzone_t *zone, qint size )
 }
 
 
-static memblock_t *SearchFree( memzone_t *zone, qint size )
+static memblock_t *SearchFree( memzone_t *zone, size_t size )
 {
 	const freeblock_t *fb;
 	memblock_t *base;
@@ -1441,7 +1441,7 @@ static memblock_t *SearchFree( memzone_t *zone, qint size )
 Z_ClearZone
 ========================
 */
-static void Z_ClearZone( memzone_t *zone, memzone_t *head, qint size, qint segnum ) {
+static void Z_ClearZone( memzone_t *zone, memzone_t *head, size_t size, qint segnum ) {
 	memblock_t	*block;
 	qint min_fragment;
 
@@ -1655,12 +1655,12 @@ Z_TagMalloc
 ================
 */
 #if defined(ZONE_DEBUG)
-void *Z_TagMallocDebug( qint size, memtag_t tag, qchar *label, qchar *file, qint line ) {
+void *Z_TagMallocDebug( size_t size, memtag_t tag, const qchar *label, const qchar *file, qint line ) {
 	qint		allocSize;
 #else
-void *Z_TagMalloc( qint size, memtag_t tag ) {
+void *Z_TagMalloc( size_t size, memtag_t tag ) {
 #endif
-	qint		extra;
+	size_t		extra;
 #if !defined(USE_MULTI_SEGMENT)
 	memblock_t	*start, *rover;
 #endif
@@ -1669,6 +1669,10 @@ void *Z_TagMalloc( qint size, memtag_t tag ) {
 
 	if ( tag == TAG_FREE ) {
 		Com_Error( ERR_FATAL, "Z_TagMalloc: tried to use with TAG_FREE" );
+	}
+
+	if ( size > INT_MAX ) {
+		Com_Error( ERR_FATAL, "Z_TagMalloc: %"PRIz"u > INT_MAX", size );
 	}
 
 	if ( tag == TAG_SMALL ) {
@@ -1712,10 +1716,10 @@ void *Z_TagMalloc( qint size, memtag_t tag ) {
 			// scanned all the way around the list
 #if defined(ZONE_DEBUG)
 			Z_LogHeap();
-			Com_Error( ERR_FATAL, "Z_Malloc: failed on allocation of %i bytes from the %s zone: %s, line: %d (%s)",
+			Com_Error( ERR_FATAL, "Z_Malloc: failed on allocation of %"PRIz"u bytes from the %s zone: %s, line: %d (%s)",
 								size, zone == smallzone ? "small" : "main", file, line, label );
 #else
-			Com_Error( ERR_FATAL, "Z_Malloc: failed on allocation of %i bytes from the %s zone",
+			Com_Error( ERR_FATAL, "Z_Malloc: failed on allocation of %"PRIz"u bytes from the %s zone",
 								size, zone == smallzone ? "small" : "main" );
 #endif
 			return NULL;
@@ -1779,9 +1783,9 @@ Z_Malloc
 ========================
 */
 #if defined(ZONE_DEBUG)
-void *Z_MallocDebug( qint size, qchar *label, qchar *file, qint line ) {
+void *Z_MallocDebug( size_t size, const qchar *label, const qchar *file, qint line ) {
 #else
-void *Z_Malloc( qint size ) {
+void *Z_Malloc( size_t size ) {
 #endif
 	void	*buf;
 
@@ -1804,11 +1808,11 @@ S_Malloc
 ========================
 */
 #if defined(ZONE_DEBUG)
-void *S_MallocDebug( qint size, qchar *label, qchar *file, qint line ) {
+void *S_MallocDebug( size_t size, const qchar *label, const qchar *file, qint line ) {
 	return Z_TagMallocDebug( size, TAG_SMALL, label, file, line );
 }
 #else
-void *S_Malloc( qint size ) {
+void *S_Malloc( size_t size ) {
 	return Z_TagMalloc( size, TAG_SMALL );
 }
 #endif
@@ -1860,7 +1864,7 @@ static void Z_LogZoneHeap( memzone_t *zone, const qchar *name ) {
 #endif
 	memblock_t	*block;
 	qchar		buf[4096];
-	qint size, allocSize, numBlocks;
+	size_t size, allocSize, numBlocks;
 	qint len;
 
 	if ( logfile == FS_INVALID_HANDLE || !FS_Initialized() )
@@ -1904,9 +1908,9 @@ static void Z_LogZoneHeap( memzone_t *zone, const qchar *name ) {
 #else
 	allocSize = numBlocks * sizeof(memblock_t); // + 32 bit alignment
 #endif
-	len = Com_sprintf( buf, sizeof( buf ), "%d %s memory in %d blocks\r\n", size, name, numBlocks );
+	len = Com_sprintf( buf, sizeof( buf ), "%"PRIz"u %s memory in %"PRIz"u blocks\r\n", size, name, numBlocks );
 	FS_Write( buf, len, logfile );
-	len = Com_sprintf( buf, sizeof( buf ), "%d %s memory overhead\r\n", size - allocSize, name );
+	len = Com_sprintf( buf, sizeof( buf ), "%"PRIz"u %s memory overhead\r\n", size - allocSize, name );
 	FS_Write( buf, len, logfile );
 	FS_Flush( logfile );
 }
@@ -2055,15 +2059,15 @@ static const qchar *tagName[ TAG_COUNT ] = {
 };
 
 typedef struct zone_stats_s {
-	qint	zoneSegments;
-	qint zoneBlocks;
-	qint	zoneBytes;
-	qint	botlibBytes;
-	qint	rendererBytes;
-	qint freeBytes;
-	qint freeBlocks;
-	qint freeSmallest;
-	qint freeLargest;
+	size_t zoneSegments;
+	size_t zoneBlocks;
+	size_t zoneBytes;
+	size_t botlibBytes;
+	size_t rendererBytes;
+	size_t freeBytes;
+	size_t freeBlocks;
+	size_t freeSmallest;
+	size_t freeLargest;
 } zone_stats_t;
 
 
@@ -2076,7 +2080,7 @@ static void Zone_Stats( const qchar *name, const memzone_t *z, qbool printDetail
 	memset( &st, 0, sizeof( st ) );
 	zone = z;
 	st.zoneSegments = 1;
-	st.freeSmallest = 0x7FFFFFFF;
+	st.freeSmallest = SIZE_MAX;
 
 	//if ( printDetails ) {
 	//	Com_Printf( "---------- %s zone segment #%i ----------\n", name, zone->segnum );
@@ -2085,7 +2089,7 @@ static void Zone_Stats( const qchar *name, const memzone_t *z, qbool printDetail
 	for ( block = zone->blocklist.next ; ; ) {
 		if ( printDetails ) {
 			qint tag = block->tag;
-			Com_Printf( "block:%p  size:%8i  tag: %s\n", (void *)block, block->size,
+			Com_Printf( "block:%p  size:%8"PRIz"u  tag: %s\n", (void *)block, block->size,
 				(unsigned)tag < TAG_COUNT ? tagName[ tag ] : va( "%i", tag ) );
 		}
 		if ( block->tag != TAG_FREE ) {
@@ -2113,7 +2117,7 @@ static void Zone_Stats( const qchar *name, const memzone_t *z, qbool printDetail
 			if ( next->size == 0 && next->id == -ZONEID && next->tag == TAG_GENERAL ) {
 				st.zoneSegments++;
 				if ( printDetails ) {
-					Com_Printf( "---------- %s zone segment #%i ----------\n", name, st.zoneSegments );
+					Com_Printf( "---------- %s zone segment #%"PRIz"u ----------\n", name, st.zoneSegments );
 				}
 				block = next->next;
 				continue;
@@ -2174,24 +2178,24 @@ static void Com_Meminfo_f( void ) {
 	Com_Printf( "\n" );
 
 	Zone_Stats( "main", mainzone, !Q_stricmp( Cmd_Argv(1), "main" ) || !Q_stricmp( Cmd_Argv(1), "all" ), &st );
-	Com_Printf( "%8i bytes total main zone\n\n", mainzone->size );
-	Com_Printf( "%8i bytes in %i main zone blocks%s\n", st.zoneBytes, st.zoneBlocks,
-		st.zoneSegments > 1 ? va( " and %i segments", st.zoneSegments ) : "" );
-	Com_Printf( "        %8i bytes in botlib\n", st.botlibBytes );
-	Com_Printf( "        %8i bytes in renderer\n", st.rendererBytes );
-	Com_Printf( "        %8i bytes in other\n", st.zoneBytes - ( st.botlibBytes + st.rendererBytes ) );
-	Com_Printf( "        %8i bytes in %i free blocks\n", st.freeBytes, st.freeBlocks );
+	Com_Printf( "%8"PRIz"u bytes total main zone\n\n", mainzone->size );
+	Com_Printf( "%8"PRIz"u bytes in %"PRIz"u main zone blocks%s\n", st.zoneBytes, st.zoneBlocks,
+		st.zoneSegments > 1 ? va( " and %"PRIz"u segments", st.zoneSegments ) : "" );
+	Com_Printf( "        %8"PRIz"u bytes in botlib\n", st.botlibBytes );
+	Com_Printf( "        %8"PRIz"u bytes in renderer\n", st.rendererBytes );
+	Com_Printf( "        %8"PRIz"u bytes in other\n", st.zoneBytes - ( st.botlibBytes + st.rendererBytes ) );
+	Com_Printf( "        %8"PRIz"u bytes in %"PRIz"u free blocks\n", st.freeBytes, st.freeBlocks );
 	if ( st.freeBlocks > 1 ) {
-		Com_Printf( "        (largest: %i bytes, smallest: %i bytes)\n\n", st.freeLargest, st.freeSmallest );
+		Com_Printf( "        (largest: %"PRIz"u bytes, smallest: %"PRIz"u bytes)\n\n", st.freeLargest, st.freeSmallest );
 	}
 
 	Zone_Stats( "small", smallzone, !Q_stricmp( Cmd_Argv(1), "small" ) || !Q_stricmp( Cmd_Argv(1), "all" ), &st );
-	Com_Printf( "%8i bytes total small zone\n\n", smallzone->size );
-	Com_Printf( "%8i bytes in %i small zone blocks%s\n", st.zoneBytes, st.zoneBlocks,
-		st.zoneSegments > 1 ? va( " and %i segments", st.zoneSegments ) : "" );
-	Com_Printf( "        %8i bytes in %i free blocks\n", st.freeBytes, st.freeBlocks );
+	Com_Printf( "%8"PRIz"u bytes total small zone\n\n", smallzone->size );
+	Com_Printf( "%8"PRIz"u bytes in %"PRIz"u small zone blocks%s\n", st.zoneBytes, st.zoneBlocks,
+		st.zoneSegments > 1 ? va( " and %"PRIz"u segments", st.zoneSegments ) : "" );
+	Com_Printf( "        %8"PRIz"u bytes in %"PRIz"u free blocks\n", st.freeBytes, st.freeBlocks );
 	if ( st.freeBlocks > 1 ) {
-		Com_Printf( "        (largest: %i bytes, smallest: %i bytes)\n\n", st.freeLargest, st.freeSmallest );
+		Com_Printf( "        (largest: %"PRIz"u bytes, smallest: %"PRIz"u bytes)\n\n", st.freeLargest, st.freeSmallest );
 	}
 }
 
@@ -2281,6 +2285,7 @@ static void Com_InitZoneMemory( void ) {
 
 	// allocate the random block zone
 	cv = Cvar_GetAndDescribe( "com_zoneMegs", XSTRING( DEF_COMZONEMEGS ), CVAR_LATCH | CVAR_ARCHIVE, "Initial amount of memory (RAM) allocated for the main block zone (in MB)." );
+	Cvar_CheckRange( cv, "1", va("%d", INT_MAX / (1024*1024)), CV_INTEGER );
 
 #if !defined(USE_MULTI_SEGMENT)
 	if ( cv->integer < DEF_COMZONEMEGS )
@@ -2397,7 +2402,7 @@ static void Com_InitHunkMemory( void ) {
 
 	// allocate the stack based hunk allocator
 	cv = Cvar_GetAndDescribe( "com_hunkMegs", XSTRING( DEF_COMHUNKMEGS ), CVAR_LATCH | CVAR_ARCHIVE, "The size of the hunk memory segment." );
-	Cvar_CheckRange(cv, XSTRING(MIN_COMHUNKMEGS), NULL, CV_INTEGER);
+	Cvar_CheckRange(cv, XSTRING(MIN_COMHUNKMEGS), va("%d", (INT_MAX-63) / (1024*1024)), CV_INTEGER);
 
 	s_hunkTotal = cv->integer * 1024 * 1024;
 
@@ -2543,15 +2548,19 @@ Allocate permanent (until the hunk is cleared) memory
 =================
 */
 #if defined(HUNK_DEBUG)
-void *Hunk_AllocDebug( qint size, ha_pref preference, qchar *label, qchar *file, qint line ) {
+void *Hunk_AllocDebug( size_t size, ha_pref preference, const qchar *label, const qchar *file, qint line ) {
 #else
-void *Hunk_Alloc( qint size, ha_pref preference ) {
+void *Hunk_Alloc( size_t size, ha_pref preference ) {
 #endif
 	void	*buf;
 
 	if ( s_hunkData == NULL)
 	{
 		Com_Error( ERR_FATAL, "Hunk_Alloc: Hunk memory system not initialized" );
+	}
+
+	if ( size > INT_MAX ) {
+		Com_Error( ERR_FATAL, "Hunk_Alloc: %"PRIz"u > INT_MAX", size );
 	}
 
 	// can't do preference if there is any temp allocated
@@ -2577,9 +2586,9 @@ void *Hunk_Alloc( qint size, ha_pref preference ) {
 		Hunk_Log();
 		Hunk_SmallLog();
 
-		Com_Error(ERR_DROP, "Hunk_Alloc failed on %i: %s, line: %d (%s)", size, file, line, label);
+		Com_Error(ERR_DROP, "Hunk_Alloc failed on %"PRIz"u: %s, line: %d (%s)", size, file, line, label);
 #else
-		Com_Error(ERR_DROP, "Hunk_Alloc failed on %i", size);
+		Com_Error(ERR_DROP, "Hunk_Alloc failed on %"PRIz"u", size);
 #endif
 	}
 
@@ -2622,7 +2631,7 @@ Multiple files can be loaded in temporary memory.
 When the files-in-use count reaches zero, all temp memory will be deleted
 =================
 */
-void *Hunk_AllocateTempMemory( qint size ) {
+void *Hunk_AllocateTempMemory( size_t size ) {
 	void		*buf;
 	hunkHeader_t	*hdr;
 
@@ -2635,12 +2644,16 @@ void *Hunk_AllocateTempMemory( qint size ) {
 		return Z_Malloc(size);
 	}
 
+	if ( size > INT_MAX ) {
+		Com_Error( ERR_FATAL, "Hunk_AllocateTempMemory: %"PRIz"u > INT_MAX", size );
+	}
+
 	Hunk_SwapBanks();
 
 	size = PAD(size, sizeof(intptr_t)) + sizeof( hunkHeader_t );
 
 	if ( hunk_temp->temp + hunk_permanent->permanent + size > s_hunkTotal ) {
-		Com_Error( ERR_DROP, "Hunk_AllocateTempMemory: failed on %i", size );
+		Com_Error( ERR_DROP, "Hunk_AllocateTempMemory: failed on %"PRIz"u", size );
 	}
 
 	if ( hunk_temp == &hunk_low ) {