+++ code/server/sv_game.c
@@ -54,6 +54,11 @@ SV_GentityNum(qint num)
 {
   sharedEntity_t *ent;
 
+  if (num < -1 || num >= MAX_GENTITIES)
+  {
+    Com_Error(ERR_DROP, "%s: bad num %d", __func__, num);
+  }
+
   ent = (sharedEntity_t *)((byte *)sv.gentities + sv.gentitySize*(num));
   return ent;
 }
@@ -209,11 +214,23 @@ SV_inPVS(const vec3_t p1, const vec3_t p2)
 
   leafnum = CM_PointLeafnum(p1);
   cluster = CM_LeafCluster(leafnum);
+
+  if (cluster < 0)
+  {
+    return qfalse;
+  }
+
   area1 = CM_LeafArea(leafnum);
   mask = CM_ClusterPVS(cluster);
 
   leafnum = CM_PointLeafnum(p2);
   cluster = CM_LeafCluster(leafnum);
+
+  if (cluster < 0)
+  {
+    return qfalse;
+  }
+
   area2 = CM_LeafArea(leafnum);
 
   if (mask && (!(mask[cluster >> 3] & (BIT(cluster & 7)))))
@@ -246,10 +263,21 @@ SV_inPVSIgnorePortals(const vec3_t p1, const vec3_t p2)
 
   leafnum = CM_PointLeafnum(p1);
   cluster = CM_LeafCluster(leafnum);
+
+  if (cluster < 0)
+  {
+    return qfalse;
+  }
+
   mask = CM_ClusterPVS(cluster);
 
   leafnum = CM_PointLeafnum(p2);
   cluster = CM_LeafCluster(leafnum);
+
+  if (cluster < 0)
+  {
+    return qfalse;
+  }
 
   if (mask && (!(mask[cluster >> 3] & (BIT(cluster & 7)))))
   {
@@ -335,31 +363,29 @@ SV_LocateGameData
 ===============
 */
 void
-SV_LocateGameData(sharedEntity_t *gEnts, qint numGEntities, qint sizeofGEntity_t, playerState_t *clients, qint sizeofGameClient)
+SV_LocateGameData(sharedEntity_t *gEnts, unsigned numGEntities, unsigned sizeofGEntity_t, playerState_t *clients, unsigned sizeofGameClient)
 {
   if (!sv.gvm->entryPoint)
   {
     if (numGEntities > MAX_GENTITIES)
     {
-      Com_Error(ERR_DROP, "%s: bad entity count %i", __func__, numGEntities);
-    }
-    else
-    {
-      if (sizeofGEntity_t > sv.gvm->exactDataLength / numGEntities)
-      {
-        Com_Error(ERR_DROP, "%s: bad entity size %i", __func__, sizeofGEntity_t);	
-      }
-      else if ((byte *)gEnts + (numGEntities * sizeofGEntity_t) > (sv.gvm->dataBase + sv.gvm->exactDataLength))
-      {
-        Com_Error(ERR_DROP, "%s: entities located out of data segment", __func__);
-      }
+      Com_Error(ERR_DROP, "%s: bad entity count %u", __func__, numGEntities);
     }
 
-    if (sizeofGameClient > sv.gvm->exactDataLength / MAX_CLIENTS)
+    if (sizeofGEntity_t < sizeof(sharedEntity_t) || sizeofGEntity_t > sv.gvm->exactDataLength / MAX_GENTITIES)
     {
-      Com_Error(ERR_DROP, "%s: bad game client size %i", __func__, sizeofGameClient);	
+      Com_Error(ERR_DROP, "%s: bad entity size %u", __func__, sizeofGEntity_t);
     }
-    else if ((byte *)clients + (sizeofGameClient * MAX_CLIENTS) > sv.gvm->dataBase + sv.gvm->exactDataLength)
+    else if ((byte *)gEnts - sv.gvm->dataBase > sv.gvm->exactDataLength - sizeofGEntity_t * MAX_GENTITIES)
+    {
+      Com_Error(ERR_DROP, "%s: entities located out of data segment", __func__);
+    }
+
+    if (sizeofGameClient < sizeof(playerState_t) || sizeofGameClient > sv.gvm->exactDataLength / MAX_CLIENTS)
+    {
+      Com_Error(ERR_DROP, "%s: bad game client size %u", __func__, sizeofGameClient);
+    }
+    else if ((byte *)clients - sv.gvm->dataBase > sv.gvm->exactDataLength - sizeofGameClient * MAX_CLIENTS)
     {
       Com_Error(ERR_DROP, "%s: clients located out of data segment", __func__);
     }
@@ -595,7 +621,7 @@ SV_GameSystemCalls(intptr_t *args)
 
     case
     G_ENTITIES_IN_BOX:
-      VM_CHECKBOUNDS(sv.gvm, args[3], args[4] * sizeof(qint));
+      VM_CHECKBOUNDS3(sv.gvm, args[3], args[4], sizeof(qint));
       return SV_AreaEntities(VMA(1), VMA(2), VMA(3), args[4]);
 
     case